Automatic Light Version 1.0
This app turns on a light when motion is detected during a user specified time period. The user can select which days this app is enabled. The user can also set a timeout period for the light. This will cause the light to turn off if the motion sensor has not detected movement in that amount of time.